+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # vNext
 
 * Fix: captureMessage defaults SentryLevel to info
+* Fix: SentryEvent.throwable returns the unwrapped throwable instead of the throwableMechanism
 
 # 4.0.5